Changed local vault encryption to use SQLCipher

+++ app/src/main/java/com/passkeyper/android/vault/local/DbHelper.java
@@ -3,22 +3,49 @@ package com.passkeyper.android.vault.local;
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
 
 import com.passkeyper.android.vaultmodel.VaultModel;
+
+import net.sqlcipher.database.SQLiteDatabase;
+import net.sqlcipher.database.SQLiteOpenHelper;
+
+import java.util.Arrays;
 
 /**
  * The helper class for the SQLite implementation of the local vault.
  */
 class DbHelper extends SQLiteOpenHelper {
 
-    DbHelper(Context context) {
+    private char[] password;
+    private boolean isClosed = false;
+
+    DbHelper(Context context, char[] password) {
         super(context, DbContract.DATABASE_NAME, null, DbContract.DATABASE_VERSION);
+        SQLiteDatabase.loadLibs(context);
+        this.password = Arrays.copyOf(password, password.length);
+    }
+
+    @Override
+    public void close() {
+        Arrays.fill(password, '\0');
+        isClosed = true;
+        super.close();
+    }
+
+    public boolean isClosed() {
+        return isClosed;
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {
+        //init the encryption
+        StringBuilder builder = new StringBuilder();
+        builder.append("PRAGMA key = '");
+        for (char c : password)
+            builder.append(c);
+        builder.append("';");
+        db.execSQL(builder.toString());
+        //build the tables
         db.execSQL(DbContract.RecordTable.SQL_CREATE_TABLE);
         db.execSQL(DbContract.SensitiveEntryTable.SQL_CREATE_TABLE);
         db.execSQL(DbContract.SecurityQuestionTable.SQL_CREATE_TABLE);
@@ -35,6 +62,25 @@ class DbHelper extends SQLiteOpenHelper {
 
     Cursor readEntry(String table, String[] projection, String recordIdColumn, long recordId) {
         return read(table, projection, recordIdColumn + " = ?", new String[]{String.valueOf(recordId)});
+    }
+
+    void resetKey(char[] key) {
+        if (isClosed())
+            throw new IllegalStateException("Database accessed after being closed");
+
+        SQLiteDatabase db = this.getWritableDatabase(password);
+
+        StringBuilder builder = new StringBuilder();
+        builder.append("PRAGMA rekey = '");
+        for (char c : key)
+            builder.append(c);
+        builder.append("';");
+
+        db.execSQL(builder.toString());
+
+        //update the password field
+        Arrays.fill(password, '\0');
+        password = Arrays.copyOf(key, key.length);
     }
 
     long save(VaultModel model, String table, String _id, ContentValues values) {
@@ -54,23 +100,33 @@ class DbHelper extends SQLiteOpenHelper {
      */
 
     private Cursor read(String table, String[] projection, String selection, String[] args) {
-        SQLiteDatabase db = this.getReadableDatabase();
+        if (isClosed())
+            throw new IllegalStateException("Database accessed after being closed");
 
-        return db.query(table, projection, selection, args, null, null, null);
+        return this.getReadableDatabase(password).query(table, projection, selection, args, null, null, null);
     }
 
     private long save(String table, ContentValues values, String selection, String[] args, boolean update) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        if (isClosed())
+            throw new IllegalStateException("Database accessed after being closed");
 
+        SQLiteDatabase db = this.getWritableDatabase(password);
+
+        long id = -1;
         if (update) db.update(table, values, selection, args);
-        else return db.insert(table, null, values);
-        return -1;
+        else id = db.insert(table, null, values);
+
+        db.close();
+        return id;
     }
 
     private void delete(String table, String selection, String[] args) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        if (isClosed())
+            throw new IllegalStateException("Database accessed after being closed");
 
+        SQLiteDatabase db = this.getWritableDatabase(password);
         db.delete(table, selection, args);
+        db.close();
     }
 
 }
